Extract common BSDF utility functions to shared module
